divide components
