+++ src/MessageBox/MessageBox.css
@@ -44,120 +44,3 @@
 	fill: white;
 	filter: drop-shadow( 2px 0px 1px rgba(0,0,0,.2) );
 }
-
-.rce-mbox-photo {
-	margin-top: -3px;
-	margin-right: -6px;
-	margin-left: -6px;
-}
-
-.rce-mbox-photo .rce-mbox-text {
-	padding: 5px 0px;
-	max-width: 300px;
-	margin: auto;
-}
-
-.rce-mbox-photo--img {
-	position: relative;
-	display: flex;
-	overflow: hidden;
-	justify-content: center;
-	border-radius: 5px;
-	max-height: 300px;
-}
-
-.rce-mbox-photo--img__block {
-	position: absolute;
-	top: 0;
-	right: 0;
-	left: 0;
-	bottom: 0;
-	background-color: rgba(0,0,0,0.5);
-	border-radius: 5px;
-	display: flex;
-}
-
-.rce-mbox-photo--img img {
-	height: 100%;
-}
-
-.rce-mbox-photo--img__block-item {
-	margin: auto;
-	cursor: pointer;
-}
-
-.rce-mbox-photo--download {
-	color: #efe;
-	display: flex;
-	justify-content: center;
-	background: none;
-	border: none;
-	font-size: 3.2em;
-	outline: none;
-	border: 1px solid #eee;
-	border-radius: 100%;
-	height: 100px;
-	width: 100px;
-}
-
-.rce-mbox-photo--download:hover {
-	opacity: .7;
-}
-
-.rce-mbox-photo--download:active {
-	opacity: .3;
-}
-
-.rce-mbox-file {
-	background: #e9e9e9;
-	display: flex;
-	border-radius: 5px;
-	margin-top: -3px;
-	margin-right: -6px;
-	margin-left: -6px;
-	align-items: center;
-	min-height: 52px;
-	max-width: 500px;
-	padding: 5px 0;
-	cursor: pointer;
-    user-select: none;
-    outline: none;
-    border:none;
-}
-
-.rce-mbox-file > * {
-	padding: 0px 10px;
-}
-
-.rce-mbox-file--icon {
-	font-size: 30px;
-	align-items: center;
-    display: flex;
-    flex-direction: column;
-}
-
-.rce-mbox-file--size {
-	font-size: 10px;
-	margin-top: 3px;
-	max-width: 52px;
-	white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.rce-mbox-file--text {
-	font-size: 13.6px;
-	white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.rce-mbox-file--buttons {
-	font-size: 30px;
-	align-items: center;
-    display: flex;
-}
-
-.rce-mbox-file--loading {
-	font-size: 15px;
-}
